split trace.go
